+++ scripts/archiso-base-211216.sh
@@ -755,9 +755,9 @@ path=/usr/bin/vivaldi-stable --app=https://meet.google.com/ --new-window
 #icon=/usr/share/icons-24x24/visualstudiocode.png
 #path=/usr/bin/vivaldi-stable --app=https://vscode.dev/ --new-window
 
-#[launcher]
-#icon=/usr/share/favicons-24x24/www.deepl.com.translator.png
-#path=/usr/bin/vivaldi-stable --app=https://www.deepl.com/translator --new-window
+[launcher]
+icon=/usr/share/favicons-24x24/www.deepl.com.translator.png
+path=/usr/bin/vivaldi-stable --app=https://www.deepl.com/translator --new-window
 
 #[launcher]
 #icon=/usr/share/favicons-24x24/www.mathcha.io.editor.png