+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 2022-01-03
 ### Fixed
 - fix btop install path
+- uncomment to add Deepl Translator to launcher
 
 ## 2022-01-02
 ### Added